Refactors Openfinex Types to Polkadex Primitives
