+++ enclave/openfinex/Cargo.toml
@@ -7,11 +7,11 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-serde = { git = "https://github.com/mesalock-linux/serde-sgx", default-features = false, features = ["mesalock_sgx"] }
 jsonwebtoken = { git = "https://github.com/mesalock-linux/jsonwebtoken-sgx.git", default-features = false, features = ["mesalock_sgx"] }
 embedded-websocket = { version = "0.8.0", default-features = false }
 sgx_tstd = { rev = "v1.1.3", git = "https://github.com/apache/teaclave-sgx-sdk.git", features = ["untrusted_fs", "net", "thread","backtrace"] }
 sgx_types = { rev = "v1.1.3", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
+polkadex-primitives = { path = "../../polkadex-primitives", default-features = false, features = ["sgx"] }
 
 [dependencies.rustls]
 git      = "https://github.com/mesalock-linux/rustls"
