+++ polkadex-primitives/Cargo.toml
@@ -10,6 +10,7 @@ edition = "2018"
 codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false, features = ["derive"] }
 frame-support = { version = "3.0.0", default-features = false, git = "https://github.com/paritytech/substrate.git" }
 sp-std = { version = "3.0.0", default-features = false, git = "https://github.com/paritytech/substrate.git" }
+serde = { git = "https://github.com/mesalock-linux/serde-sgx", default-features = false, features = ["mesalock_sgx"] }
 
 [dependencies.sp-core]
 git = "https://github.com/paritytech/substrate.git"
